Reorganize the folder structure
