+++ src/components/Airship.tsx
@@ -1,45 +1,7 @@
 import * as React from 'react'
 
-import { Wrapper } from './wrapper/Wrapper'
-
-/**
- * Control panel for managing a component inside an airship.
- */
-export interface AirshipBridge<T> {
-  // Use these to pass values to the outside world:
-  resolve(value: T | PromiseLike<T>): void
-  reject(error: Error): void
-
-  // Unmounts the component:
-  remove(): void
-
-  // Runs a callback when the result promise settles.
-  // Useful for starting exit animations:
-  onResult(callback: () => unknown): void
-}
-
-interface Props {
-  children?: React.ReactNode
-  avoidAndroidKeyboard?: boolean
-  statusBarTranslucent?: boolean
-}
-
-/**
- * Renders a component to place inside the airship.
- */
-type AirshipRender<T> = (bridge: AirshipBridge<T>) => React.ReactNode
-
-/**
- * The airship itself is a component you should mount after your main
- * scene or router.
- *
- * It has a static method anyone can call to display components.
- * The method returns a promise, which the component can use to pass values
- * to the outside world.
- */
-export interface Airship extends React.ComponentClass<Props> {
-  show<T>(render: AirshipRender<T>): Promise<T>
-}
+import { Airship, AirshipBridge, AirshipProps, AirshipRender } from '../types'
+import { Wrapper } from './Wrapper'
 
 /**
  * Constructs an Airship component.
@@ -54,8 +16,8 @@ export function makeAirship(): Airship {
     for (const container of allContainers) container.forceUpdate()
   }
 
-  class AirshipContainer extends React.Component<Props> {
-    constructor(props: Props) {
+  class AirshipContainer extends React.Component<AirshipProps> {
+    constructor(props: AirshipProps) {
       super(props)
       allContainers.push(this)
     }
