+++ src/components/AirshipModal.tsx
@@ -7,8 +7,8 @@ import {
   ViewStyle
 } from 'react-native'
 
-import { AirshipBridge } from '../Airship'
-import { unpackEdges } from '../utils'
+import { AirshipBridge } from '../types'
+import { unpackEdges } from '../util/edges'
 
 export interface AirshipModalProps<T = unknown> {
   bridge: AirshipBridge<T>
