+++ src/index.ts
@@ -1,4 +1,8 @@
-export * from './Airship'
-export * from './components/AirshipDropdown'
-export * from './components/AirshipModal'
-export * from './components/AirshipToast'
+export { makeAirship } from './components/Airship'
+export {
+  AirshipDropdown,
+  AirshipDropdownProps
+} from './components/AirshipDropdown'
+export { AirshipModal, AirshipModalProps } from './components/AirshipModal'
+export { AirshipToast, AirshipToastProps } from './components/AirshipToast'
+export { Airship, AirshipBridge } from './types'
