+++ tsconfig.json
@@ -14,5 +14,5 @@
 
     "strict": true
   },
-  "exclude": ["node_modules", "lib"]
+  "exclude": ["AirshipDemo/node_modules", "lib"]
 }
